Tidy up the test suite as much as possible.
