+++ src/test/scala/sss/db/SqlInterpolatorSpec.scala
@@ -4,6 +4,7 @@ import java.util.Date
 
 import org.scalatest._
 
+@DoNotDiscover
 class SqlInterpolatorSpec extends FlatSpec with Matchers with BeforeAndAfter {
 
   it should " interpolate the string replacing the params with ? for prepared statements." in {
@@ -11,8 +12,8 @@ class SqlInterpolatorSpec extends FlatSpec with Matchers with BeforeAndAfter {
     val c = new Date()
 
     val (goodSql, params)  = ps"SELECT * from x where a = $b AND b = ${c}"
-    assert(goodSql == "SELECT * from x where a = ? AND b = ?")
-    assert(params == Seq(56, c))
+    assert(goodSql === "SELECT * from x where a = ? AND b = ?")
+    assert(params === Seq(56, c))
 
   }
 }
